# docs/api/swagger.yaml
basePath: /api/v1
...
host: localhost:8080
info:
  contact:
    email: hello@goatflow.io
    name: GoatFlow Support
    url: https://goatflow.io/support
  description: GoatFlow Ticket System REST API
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://goatflow.io/terms/
  title: GoatFlow API
  version: "1.0"
paths:
...
  /tickets/{id}:
    delete:
      consumes:
      - application/json
      description: Archive a ticket (soft delete - tickets are never hard deleted)
      parameters:
      - description: Ticket ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Ticket archived
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Ticket not found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete (archive) ticket
      tags:
      - Tickets
    get:
      consumes:
      - application/json
      description: Retrieve a single ticket by its ID with optional related data
      parameters:
      - description: Ticket ID
        in: path
        name: id
        required: true
        type: integer
      - description: 'Include related data (comma-separated: articles, customer, queue)'
        in: query
        name: include
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Ticket details
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Ticket not found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get ticket by ID
      tags:
      - Tickets
    put:
      consumes:
      - application/json
      description: Update an existing ticket's properties
      parameters:
      - description: Ticket ID
        in: path
        name: id
        required: true
        type: integer
      - description: Ticket update data (title, queue_id, priority_id, state_id, etc.)
        in: body
        name: ticket
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Updated ticket
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Ticket not found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update ticket
      tags:
      - Tickets
...
securityDefinitions:
  BearerAuth:
    description: API token (Bearer gf_xxx) or JWT token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"