# internal/api/ticket_get_handler.go
package api

import (
	"database/sql"
	"fmt"
	"log"
	"net/http"
	"strconv"
	"time"

	"github.com/gin-gonic/gin"

	"github.com/goatkit/goatflow/internal/platform/database"
	"github.com/goatkit/goatflow/internal/services"
)

// HandleGetTicketAPI handles GET /api/v1/tickets/:id.
//
//	@Summary		Get ticket by ID
//	@Description	Retrieve a single ticket by its ID with optional related data
//	@Tags			Tickets
//	@Accept			json
//	@Produce		json
//	@Param			id		path		int		true	"Ticket ID"
//	@Param			include	query		string	false	"Include related data (comma-separated: articles, customer, queue)"
//	@Success		200		{object}	map[string]interface{}	"Ticket details"
//	@Failure		401		{object}	map[string]interface{}	"Unauthorized"
//	@Failure		404		{object}	map[string]interface{}	"Ticket not found"
//	@Security		BearerAuth
//	@Router			/tickets/{id} [get]
func HandleGetTicketAPI(c *gin.Context) {
	ticketIDStr := c.Param("id")
	ticketID, err := strconv.ParseInt(ticketIDStr, 10, 64)
	if err != nil {
		c.JSON(http.StatusBadRequest, gin.H{
			"success": false,
			"error":   "Invalid ticket ID",
		})
		return
	}

	if _, exists := c.Get("user_id"); !exists {
		if _, authExists := c.Get("is_authenticated"); !authExists {
			c.JSON(http.StatusUnauthorized, gin.H{
				"success": false,
				"error":   "Authentication required",
			})
			return
		}
	}

	db, err := database.GetDB()
	if err != nil || db == nil {
		log.Printf("HandleGetTicketAPI: database unavailable: %v", err)
		c.JSON(http.StatusInternalServerError, gin.H{
			"success": false,
			"error":   "Database unavailable",
		})
		return
	}

	// Query for ticket details with related information (GoatFlow/OTRS schema)
	typeSelect := fmt.Sprintf("%s AS type_id", database.QualifiedTicketTypeColumn("t"))
	query := database.ConvertPlaceholders(fmt.Sprintf(`
		SELECT 
			t.id,
			t.tn as ticket_number,
			t.title,
			t.ticket_state_id,
			ts.name as state_name,
			t.ticket_priority_id,
			tp.name as priority_name,
			t.queue_id,
			q.name as queue_name,
			%s,
			t.customer_id,
			t.customer_user_id,
			t.user_id as owner_user_id,
			t.responsible_user_id,
			t.create_time,
			t.change_time
		FROM ticket t
		LEFT JOIN ticket_state ts ON t.ticket_state_id = ts.id
		LEFT JOIN ticket_priority tp ON t.ticket_priority_id = tp.id
		LEFT JOIN queue q ON t.queue_id = q.id
		WHERE t.id = ?
	`, typeSelect)) //nolint:gk-sql-sprintf // hardcoded column fragments; user values bound via ?

	var ticket struct {
		ID                int64          `json:"id"`
		TicketNumber      string         `json:"ticket_number"`
		Title             string         `json:"title"`
		StateID           int            `json:"state_id"`
		StateName         sql.NullString `json:"-"`
		PriorityID        int            `json:"priority_id"`
		PriorityName      sql.NullString `json:"-"`
		QueueID           int            `json:"queue_id"`
		QueueName         sql.NullString `json:"-"`
		TypeID            sql.NullInt32  `json:"-"`
		CustomerID        sql.NullString `json:"-"`
		CustomerUserID    sql.NullString `json:"-"`
		OwnerUserID       int            `json:"owner_user_id"`
		ResponsibleUserID sql.NullInt32  `json:"-"`
		CreateTime        time.Time      `json:"create_time"`
		ChangeTime        time.Time      `json:"change_time"`
	}

	err = db.QueryRow(query, ticketID).Scan(
		&ticket.ID,
		&ticket.TicketNumber,
		&ticket.Title,
		&ticket.StateID,
		&ticket.StateName,
		&ticket.PriorityID,
		&ticket.PriorityName,
		&ticket.QueueID,
		&ticket.QueueName,
		&ticket.TypeID,
		&ticket.CustomerID,
		&ticket.CustomerUserID,
		&ticket.OwnerUserID,
		&ticket.ResponsibleUserID,
		&ticket.CreateTime,
		&ticket.ChangeTime,
	)

	if err == sql.ErrNoRows {
		c.JSON(http.StatusNotFound, gin.H{
			"success": false,
			"error":   "Ticket not found",
		})
		return
	}

	if err != nil {
		c.JSON(http.StatusInternalServerError, gin.H{
			"success": false,
			"error":   "Failed to retrieve ticket",
		})
		return
	}

	// Check queue access permissions for agents
	// Customers have separate isolation logic (company isolation)
	userRole, _ := c.Get("user_role")
	isCustomer := userRole == "Customer"
	if !isCustomer {
		userID := 0
		if ctxUserID, exists := c.Get("user_id"); exists {
			switch v := ctxUserID.(type) {
			case int:
				userID = v
			case int64:
				userID = int(v)
			case uint:
				userID = int(v)
			}
		}
		if userID <= 0 {
			c.JSON(http.StatusUnauthorized, gin.H{
				"success": false,
				"error":   "Authentication required",
			})
			return
		}

		permSvc := services.NewPermissionService(db)
		canRead, err := permSvc.CanReadQueue(userID, ticket.QueueID)
		if err != nil {
			c.JSON(http.StatusInternalServerError, gin.H{
				"success": false,
				"error":   "Failed to check permissions",
			})
			return
		}
		// Security: return 404 (not 403) to avoid revealing ticket existence
		if !canRead {
			c.JSON(http.StatusNotFound, gin.H{
				"success": false,
				"error":   "Ticket not found",
			})
			return
		}
	} else {
		// Customer company isolation: customers can only see tickets from their company
		customerLogin, _ := c.Get("customer_login")
		if loginStr, ok := customerLogin.(string); ok && loginStr != "" {
			// Look up customer's company
			var customerCompany string
			err := db.QueryRow(database.ConvertPlaceholders(
				"SELECT customer_id FROM customer_user WHERE login = ?",
			), loginStr).Scan(&customerCompany)

			if err != nil {
				c.JSON(http.StatusForbidden, gin.H{
					"success": false,
					"error":   "Access denied - customer not found",
				})
				return
			}

			// Check if ticket belongs to customer's company
			// Security: return 404 (not 403) to avoid revealing ticket existence
			if !ticket.CustomerID.Valid || ticket.CustomerID.String != customerCompany {
				c.JSON(http.StatusNotFound, gin.H{
					"success": false,
					"error":   "Ticket not found",
				})
				return
			}
		}
	}

	// Build response with formatted data
	response := gin.H{
		"id":            ticket.ID,
		"ticket_number": ticket.TicketNumber,
		"title":         ticket.Title,
		"state_id":      ticket.StateID,
		"state":         ticket.StateName.String,
		"priority_id":   ticket.PriorityID,
		"priority":      ticket.PriorityName.String,
		"queue_id":      ticket.QueueID,
		"queue":         ticket.QueueName.String,
		"owner_user_id": ticket.OwnerUserID,
		"create_time":   ticket.CreateTime.Format(time.RFC3339),
		"change_time":   ticket.ChangeTime.Format(time.RFC3339),
	}

	// Add optional fields if they have values
	if ticket.TypeID.Valid {
		response["type_id"] = ticket.TypeID.Int32
	}
	if ticket.CustomerID.Valid {
		response["customer_id"] = ticket.CustomerID.String
	}
	if ticket.CustomerUserID.Valid {
		response["customer_user_id"] = ticket.CustomerUserID.String
	}
	if ticket.ResponsibleUserID.Valid {
		response["responsible_user_id"] = ticket.ResponsibleUserID.Int32
	}

	// Get article count
	var articleCount int
	if err := db.QueryRow(database.ConvertPlaceholders(`
		SELECT COUNT(*) FROM article WHERE ticket_id = ?
	`), ticketID).Scan(&articleCount); err != nil {
		log.Printf("HandleGetTicketAPI: article count for ticket %d: %v", ticketID, err)
		c.JSON(http.StatusInternalServerError, gin.H{
			"success": false,
			"error":   "Failed to retrieve ticket",
		})
		return
	}
	response["article_count"] = articleCount

	c.JSON(http.StatusOK, gin.H{
		"success": true,
		"data":    response,
	})
}